The final push for mysql integration. Should be working? Since there's a lot of dependence on the day of the week, it's not finished testing.
We'll see what happens over the course of the next week.

Fully (as to my knowledge, I may have missed something) brought over Tonny's C#
code, and translated it to Javascript.

Added the ability to schedule, using Cron. Right now it's used to tell the server
 at the start of each day to generate the list of people for that day.
The reason for this is mainly because it would be too slow to have it done on demand
since Javascript doesn't allow for any actual threading. Apparently, according to
Tonny, there is a table that should be reset once a year. This will be done with Cron
to make it easier on us.

At some point in the future I will come back and add some explanatory comments
in the actual code.